refactoring filestrucutre
