+++ mc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,4 +1,4 @@
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     // Base type for our SyntaxNodes
     public abstract class SyntaxNode
