+++ mc/CodeAnalysis/Syntax/SyntaxTree.cs
@@ -1,4 +1,4 @@
-namespace Minsk.CodeAnalysis
+namespace Minsk.CodeAnalysis.Syntax
 {
     // The main Interface/API to interact with to start the parsing process
     public sealed class SyntaxTree
